pyproject.toml
[tool.poetry]
name = "whoop-plus"
version = "0.1.0"
description = "Track long term trends in your health using your WHOOP device. Get your first month with WHOOP free at https://join.whoop.com/46A8F8."
authors = ["ericaleman <mail@ericaleman.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.21.0"
Authlib = "^1.2.0"
requests = "^2.28.1"
backoff = "^2.2.1"
ratelimit = "^2.2.1"
pandas = "^1.5.2"
pandasql = "^0.7.3"
click = "^8.1.3"
matplotlib = "^3.6.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
